fix: restore typescript 5.6 compatibility (#301)
This PR restores compatibility with TypeScript 5.6. It also updates the
test suite to be more resilient and explicitly tests against TypeScript
5.5 and 5.6.

I was surprised that TypeScript would break this without releasing a
major version. However, if you look at the `getOriginalSourceFile`
method we were using, it was marked as `@internal` in the JSDoc:
https://github.com/microsoft/TypeScript/blob/c8a7d589e647e19c94150d9892909f3aa93e48eb/src/compiler/utilities.ts#L5416-L5419

Because it was marked as internal, they probably reserved the right to
remove this method without a breaking version bump. The other PR to fix
this issue (#270) suggested using `getSourceFileOfNode`, but that's also
marked as `@internal`:
https://github.com/microsoft/TypeScript/blob/11b2930fa2c9f73b0ffb725a9715b8d3c4121bbc/src/compiler/utilities.ts#L965-L975.

Therefore, instead, I updated this code to use `getParseTreeNode` and
`isSourceFile`, which are both exposed publicly and not marked as
`internal`:
https://github.com/microsoft/TypeScript/blob/11b2930fa2c9f73b0ffb725a9715b8d3c4121bbc/src/compiler/utilitiesPublic.ts#L803-L809
/
https://github.com/microsoft/TypeScript/blob/11b2930fa2c9f73b0ffb725a9715b8d3c4121bbc/src/compiler/factory/nodeTests.ts#L1013-L1016

By using these methods, we should be more resilient to internal
refactoring.

This PR supersedes #270.
Fixes #266